pyproject.toml
[project]
name = "game"
version = "0.1.0"
description = "hacker-ish platformer built with pygame-ce"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.6",
]
